+++ src/main/java/walkingkooka/spreadsheet/dominokit/fetcher/SpreadsheetImporterFetcherWatcher.java
@@ -19,7 +19,7 @@ package walkingkooka.spreadsheet.dominokit.fetcher;
 
 import walkingkooka.spreadsheet.dominokit.AppContext;
 import walkingkooka.spreadsheet.importer.SpreadsheetImporter;
-import walkingkooka.spreadsheet.importer.SpreadsheetImporterInfoSet;
+import walkingkooka.spreadsheet.importer.provider.SpreadsheetImporterInfoSet;
 
 /**
  * A watcher that receives all {@link SpreadsheetImporter} response events.
