+++ src/main/java/walkingkooka/spreadsheet/dominokit/fetcher/SpreadsheetImporterFetcherWatchers.java
@@ -18,7 +18,7 @@
 package walkingkooka.spreadsheet.dominokit.fetcher;
 
 import walkingkooka.spreadsheet.dominokit.AppContext;
-import walkingkooka.spreadsheet.importer.SpreadsheetImporterInfoSet;
+import walkingkooka.spreadsheet.importer.provider.SpreadsheetImporterInfoSet;
 
 public final class SpreadsheetImporterFetcherWatchers extends FetcherWatchers<SpreadsheetImporterFetcherWatcher>
     implements SpreadsheetImporterFetcherWatcher {
